Cambio tipo de dato cantidad a double(Cambio en la base de datos tambien). Arreglo algunos errores

+++ app/src/main/java/com/example/superlista/data/SuperListaDbManager.java
@@ -461,7 +461,7 @@ public class SuperListaDbManager {
         }
     }
 
-    public void updateCantidadProductoLista(ProductoPorLista prod, int cantidad){
+    public void updateCantidadProductoLista(ProductoPorLista prod, double cantidad){
         try {
             prod.setCantidad(cantidad);
             getHelper().getProductoPorListaDao().update(prod);
